+++ tactics_project/src/com/marklynch/utils/ArrayUtils.java
@@ -4,7 +4,7 @@ import com.marklynch.level.squares.Square;
 
 public class ArrayUtils {
 
-	public static boolean inBounds(Square[][] array, float index1, float index2) {
-		return (index1 >= 0) && (index1 < array.length) && (index2 >= 0) && (index2 < array[(int) index1].length);
+	public static boolean inBounds(Square[][] array, int index1, int index2) {
+		return (index1 >= 0) && (index1 < array.length) && (index2 >= 0) && (index2 < array[index1].length);
 	}
 }
